refactor web to @frctl/web
