+++ packages/fractal/package.json
@@ -18,31 +18,25 @@
     "@frctl/handlebars": "^1.2.0",
     "@frctl/mandelbrot": "^1.4.0",
     "@frctl/core": "^0.0.0",
+    "@frctl/web": "^0.0.0",
     "anymatch": "^2.0.0",
     "bluebird": "^3.5.2",
-    "browser-sync": "^2.26.4",
     "chalk": "^1.1.3",
     "chokidar": "^3.3.1",
     "cli-table3": "^0.5.0",
     "co": "^4.6.0",
     "columnify": "^1.5.4",
     "execa": "^4.0.1",
-    "express": "^4.14.0",
     "fs-extra": "^9.0.0",
     "gray-matter": "^2.0.2",
     "handlebars": "^4.0.13",
     "inquirer": "^1.1.2",
-    "js-yaml": "^3.13.1",
     "liftoff": "^2.3.0",
     "lodash": "^4.17.11",
     "log-update": "^1.0.2",
     "mime": "^1.3.4",
-    "nunjucks": "^3.2.1",
-    "path-to-regexp": "^1.5.3",
-    "portscanner": "^1.0.0",
     "require-all": "^2.0.0",
     "semver": "^5.3.0",
-    "throat": "^3.0.0",
     "update-notifier": "^2.0.0",
     "vinyl": "^1.2.0",
     "vorpal": "^1.12.0"
