+++ packages/fractal/test/web.js
@@ -4,7 +4,7 @@ const chai = require('chai');
 const expect = chai.expect;
 
 const app = require('../src/fractal')();
-const Web = require('../src/web');
+const Web = require('@frctl/web').Web;
 
 describe('Web', function () {
     let web;
