+++ packages/mandelbrot/package.json
@@ -23,7 +23,6 @@
     "@frctl/fractal": ">= 1.1.0-alpha.0 < 2"
   },
   "devDependencies": {
-    "@frctl/fractal": "^1.3.0",
     "babel-eslint": "^10.1.0",
     "babel-preset-es2015": "^6.5.0",
     "babelify": "^7.2.0",
@@ -52,6 +51,7 @@
     "watchify": "^3.7.0"
   },
   "dependencies": {
+    "@frctl/web": "^0.0.0",
     "js-beautify": "^1.6.2",
     "lodash": "^4.17.11"
   }
