+++ packages/web/src/engine/globals/static.js
