+++ packages/web/src/error.js
