+++ packages/web/src/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+    Builder: require('./builder'),
+    Engine: require('./engine'),
+    Server: require('./server'),
+    Theme: require('./theme'),
+    Web: require('./web'),
+    WebError: require('./error'),
+};
